changed the path of test_my_local_db.rb
