+++ test/run_test.rb
@@ -1,5 +1,5 @@
 base_dir = File.expand_path(File.join(File.dirname(__FILE__), '..'))
-test_dir = File.join(base_dir, 'test')
+test_dir = File.join(base_dir, 'v2/test')
 
 require 'test/unit'
